+++ Stitch/Graph/ViewModel/StitchDocumentViewModel.swift
@@ -148,7 +148,7 @@ final class StitchDocumentViewModel: Sendable {
     @MainActor var stitchAITrainingTip = StitchAITrainingTip()
     
     // AI Examples overlay panel state
-    @MainActor var showAIExamples = false
+    @MainActor var showAITrainingExamplesOverlay = false
     
 #if !targetEnvironment(macCatalyst)
     @MainActor var selectedTab = ProjectTab.patch
